Start implementing Arduino endpoint side.
Also add stub files for the rest of the files that are projected to be
in this project and add a few error codes to the spec.

Note that several corners were cut, including not bothering with the
implementations of the send functions for message types that shouldn't
be sent from this endpoint, not bothering to implement receiving message
types that we should not receive at this endpoint (an
INVALID_MESSAGE_TYPE is sent back), and a few function implementations
that are definitely suboptimal.

More importantly, be aware that this code is CURRENTLY UNTESTED.