Change abstraction for receive hibike message functions.
We previously planned to have to check the serial port to see if any
data was available before unmarshalling a hibike message. This ended
up being a bit silly abstraction-wise since use of Arduino.Serial was
not used anywhere but there. Change how this works and the inline
documentation.

This change may have to be reverted eventually in case the extra
function calls in the loop body end up taking up too much overhead,
but it probably doesn't really matter.