Clean up code to be in a working, semi-readable state.
Pre-optimized Hibike runs comfortably at a decent speed of 50Hz using
spoofed data. No stress testing has been attempted (50Hz just happened
to be the first speed tested), so it may be possible to have updates run
faster. Note that these speeds are using faked sensor data, but the
methods used to fake data should not be very different from reading a
sensor in practice.

There are still, of course, a large number of improvements that can be
made, including but not limited to:
  - better error handling
  - better debug logs
  - finishing implementing the message types deferred for now
  - message retries in certain error cases
  - multithreaded frank-side code
  - proper, automagic enumeration for Arduinos