Start work rewriting hibike.
Now that we have more reason to suspect that the hibike protocol works
and is somewhat viable (a bit more stress testing since last time has
shown that we can send messages at ~100Hz with no problems), it's time
to start making the code NOT horrendous!

Start (more or less) rewriting the entire protocol from scratch to be
object oriented, as well as actually in C++, on the Arduino side. The
goal is to make the python implementation on the Beaglebone end as
similar as possible for ease of maintenance. Additionally, new staff
will be able to simple shift to a different syntax instead of having to
pick up a new programming paradigm.