Reimplement hibike_message.h/cpp to be object oriented.
The decision to have the Arduino endpoints of the hibike protocol be
object oriented was made so that the programming paradigms of both endpoints
of the hibike protocol are the same.

Because of the discovery that it's not too hard to write standard C++
for Arduino and ignore the god-awful Arduino IDE entirely, several other
changes were made.
 - switch to a more standard project directory structure and naming
   conventions.
   - rename ino/limit-switch.ino -> src/limit_switch.cpp, etc.
   - move hibike.py into src/
 - reintroduce use of typed enums as opposed to define statements.

Note that these changes (use of features present beginning in C++11 and
the cpp standard library) make the hibike codebase incompatible with the
Arduino IDE (see https://github.com/ladislas/Bare-Arduino-Project).