Start python implementation of HibikeMessage class and subclasses.
Using python abstract base classes should probably be considered, but
their behavior seems kinda strange compared to standard C++ notion of a
base class so it may not be worth the confusion.

Note that the __calculateChecksum() and send() methods need to be
implemented for each HibikeMessage subclass as well as the (not class
related) receiveHibikeMessage() function.