Import frankfurter.
It was necessary to change a few of the paths since everything is
relative to a different repo root dir now. These changes haven't been
tested yet, but we're going ahead with merging this anyway and will just
keep the original frankfurter repo around until things are tested.

Also, we've temporarily changed the name of the communal chromebook
project to chrommunal until we think of something better.