Fix axiom permission issues.
The fix ended up being a relatively simple config file change that just
forces the user accessing the samba share to be vagrant. This would be
relatively insecure in any context aside from running a network drive on
a local VM, but it should be okay for our uses.

Also, removed the enum34 python dependency and added a bit more
documentation.