changed to multiprocessing
Rewrote the code to use multiprocessing instead of threading. This is
necessary so that we can kill nonresponsive student code (ex. infinite
loop) via timers and signals. In a threaded model, the generated signal
may be caught by any other thread.

Turns out that multiprocessing actually makes the code a little nicer. A
badThingsQueue is now used to communicate badThings, and it comes with
its own locks and stuff internally implemented; all we need to do is put
from one process, and get(block=true) from another.

The current student code modifies its own entirely local copy of state.
It will probably be until I can write a studentApi that I can fully test
the communication of state across the process barrier. However, the fact
that I can send traceback info already is a good sign that it should not
be too difficult.