stateManger to share state
Now requires Python3 to function due a bug where passing a Connection
over a Pipe/Queue will crash.

stateManager has been added to manage robot state between processes. On
construction, the object takes an inputQueue. This queue is also given
to any processes that runtime spawns. stateManager listens on this
inputQueue for lists of strings that contain instructions, such as
["reset"] or ["add", pipe].

Before being able to reply to any process's request for state data, that
process's pipe must be added to stateManger by runtime sending the
["add", pipe] message. If the process's request for information
(currently just a ["hello"] message) gets put on the Queue before the
"add", then the stateManager will crash because it does not yet know
about that process. Thus, each process must wait for a "ready" message
that the stateManager will send it upon adding a new pipe.

initialization of robot state has been moved from runtime.py to the
StateManager class.

in runtime.py: badThingsQueue has been moved from global scope to the
stack within the runtime() method.

the main runtime loop will only restart the student code if the student
code throws an error. This is accomplished by making the BadThing store
its event type and whether or not its __str__ method returns a stack
trace or the data the instance was initialized with.