refactoring
Refactoring the code to be cleaner.

Enum class is used to list the BAD_EVENTS.
Enum class is used to list the process names. Moved to runtimeUtil.
Enum class is used to list StateManager commands
Enum class is used to list Runtime Info like debug_string

Created higher order function processFactory in runtime to de-duplicate
creation of new processes. Now, a simple spawnProcess(processName,
newProcessFunction) call will create a new process and hook it up with
all the queues and pipes to runtime and the state manager.