implemented studentAPI
studentAPI.py is now a useful file!

This commit adds a layer of indirection in between studentCode and
runtime. This will allow us to sanitize any studentCode requests and
react appropriately. For exmample, attempting to set the team the
student robot is on should probably not be allowed.

stduentAPI provides a Robot class that is meant to represent the state
of the robot *to the student*. Robot is designed to abstract away all
the nastiness of dealing with pipes and queues that are necessary for
communicating with the state manager. This is done by creating the Robot
object in runtime, then setting the 'Robot' attribute of the imported studentCode module to this newly created object.

Importing of studentCode has been moved to the runStudentCode process.

Various TODO's have been added describing the next steps to take.