implements feedback from code review
This commit implements the feedback from comment in pull request.

Refactor of globalBadThing variable to newBadThing. Initialization is
removed, because we wont inspect a badThing until a badThing is in the
badThingQueue.

import studentCode is now in a timeout.

changed branch format in README from devel_{your name} to devel/{your
name}

Added newlines to end of files

Prints message when terminating runtime as part of normal operation.

Only runs runtime() if __name__ == "__main__"

Added TODO comments to address other comments.