[RUNTIME] stateManager sends on newly created pipe (#19)
When a new process is added and it's pipes hooked up to stateManager,
stateManager responds by sending the string "ready" on this pipe. The
newly spawned process should wait to receive this string before
continuing use of the pipe.

This change was so that projects external to runtime (e.x. Hibike) can
expect a standard response on their pipes, and not need to import
runtimeUtil.py.

A minor refactoring was done.
The Enum class RUNTIME_INFO has been renamed to the more representativ
RUNTIME_CONFIG.
The SM_COMMANDS.READY Enum has been removed and moved to
RUNTIME_CONFIG.PIPE_READY.