[RUNTIME] stateManager bugfix (#27)
stateManager was responding on a pipe to studentAPI, when the studentAPI
was not expecting it. This caused a backlog of read/writes on the pipes,
leading to weird behaviour.

This commit fixes this bug by making studentAPI read from the pipe and
return the value, just like stateManager is doing.