[RUNTIME] Beaglebone Compatability Fixes (#67)
This corrects some halting that was occuring when running runtime.py on
the Beaglebone due to it having only 1 CPU. This also fixes a bug that
caused runStudentCode to fail when the studentCode ran longer than the
set frequency for the code to rerun. Also adds files "disable_CPUs.sh"
and "enable_CPUs.sh" for simulating the Beaglebone when testing.

In order to check that your code would run on the Beaglebone, run
disable_CPUs.sh in order to leave only one CPU running on your machine
to simulate the Beaglebone or run it on the Beaglebone. You can then
reenable them by running enable_CPUs.sh whenever you want. This only
works on Linux as far as I know.
-Carlos

For some reason, multiprocessing Queues break on a single core when the
student code is sent a SIGKILL signal. This leads to all of runtime
hanging indefinitely on a corrupted badThingsQueue.

This commit works around this problem by sending the SIGTERM signal
first. runStudentCode() defines and registers a signal handler to catch
this, and break out of the loop executing studentCode.main.

In the case that a SIGKILL is necessary, we can no longer guarantee the
integrity of runtime's interprocess communication objects, and so we
must restart everything.
THIS IS NOT YET IMPLEMENTED.

Two new tests have been added that test infinite loops in the setup and
main functions, respectively. These tests show that runtime can
successfully restart studentCode with only a single core.
-Yizhe