[HIBIKE] handle errors when opening serial port in hibike_process.py (#123)
When hibike_process.py failed to open a serial port, for example when one of the virtual devices in virtual_devices.txt turned out to be invalid, it would crash the whole process. This fix is a hack that instead only crashes the individual worker threads corresponding to the offending serial port. This fix should be replaced with an overall better management of worker threads in a later pull request.