[DevOps] Set up Travis to build dawn and runtime (#131)
By default, travis will run dawn tests, runtime tests, and package
a runtime update tarball (but not do anything with it).

If a commit is tagged starting with dawn/ or test/dawn/, it will
additionally build Dawn binaries and upload them to GitHub releases.

If a commit is tagged starting with runtime/ or test/runtime, it
will additionally upload the runtime update tarball to GitHub
releases.