[HIBIKE] Implement a disable packet & function (#196)
Implement the README's Example Device.  Works w/ Hibike.
Re-tool the example device to impliment more datatypes, so that it
meshes with the example device in the .json and README files.

Discovered that Arduino's doubles are the same size as floats.  Asking
for a double in the .json will wait for 8 bytes, but the arduino will
only ever send 4.
Included a helpful warning about this specific issue.

Fix json file so it works with the example device.

* Update Hibike Tester to actually work with Hibike Process

Had to rename several messages, and reformat their payloads to bring in
line with RUNTIME_INTEGRATION.md

* Impliment commandable Disable in Hibike Python & Embedded.

Impliment a commandable disable using a special disable packet.

Hibike generates a new packet and sends it out to all connected devices,
commanding a disable.

Each Embedded Hibike Device must impliment a disable() function, which
is called by hibike when it recieves a hibike function.

This functionality has been tested to work with a special variety of
example_device and hibike_tester.

This commit will break all embdeed hibike devices for the time being,
untill they can impliment a disable()

* Update RUNTIME_Integration doucmentation to match what previous commits had.

* Every smart sensor now has a skeleton disable()

SO that they'll compile.

* Style fixes.

* more newlines at end of files.

* More style fixes.

* More eol fixes