[SHEPHERD] Start Integration (#473)
This PR is Sam and I creating "Dummy" versions of the Scoreboard, Control UI, and Sensors in order to test the different components of Shepherd that we currently have. We made minimal changes to existing pieces of code (just enough to get it to work) and worked out the state machine for field control (and started writing the main file).